requirements.txt
discord.py
dnspython==1.16.0
PyNaCl==1.3.0